Create a wrapper service for all geojson output handling (similar to other services)
