+++ application/src/main/resources/application.properties
@@ -48,7 +48,8 @@ analytics.datasource.flyway.locations=classpath:db/migration/analytics,classpath
 
 geojson.enabled=true
 geojson.targetRestEndpointUrl=http://localhost:12345
-geojson.requestTimeout=2000
+geojson.requestTimeoutMs=2000
+geojson.jobIncludeList=job1
 
 # logging.level.org.springframework.security=DEBUG
 # logging.level.org.springframework.web=DEBUG
