+++ service/src/main/java/de/starwit/service/geojson/GeoJsonService.java
@@ -1,64 +1,42 @@
 package de.starwit.service.geojson;
 
-import java.net.URI;
-import java.time.Duration;
+import java.util.List;
 
-import org.geojson.GeoJsonObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.MediaType;
-import org.springframework.http.client.JdkClientHttpRequestFactory;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestClient;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import de.starwit.service.jobs.AreaOccupancyObservation;
+import de.starwit.service.jobs.LineCrossingObservation;
 
 @Service
 public class GeoJsonService {
-
-    private static final Logger log = LoggerFactory.getLogger(GeoJsonService.class.getName());
-
-    private final RestClient restClient;
-    private final ObjectMapper objectMapper;
-
-    @Value("${geojson.enabled:false}")
-    private boolean geoJsonEnabled;
-
-    @Value("${geojson.targetRestEndpointUrl:http://localhost:12345}")
-    private URI GEO_JSON_URL;
-
-    @Value("${geojson.requestTimeoutMs:2000}")
-    private int requestTimeoutMs;
-
-    public GeoJsonService() {
-        this.objectMapper = new ObjectMapper();
-
-        JdkClientHttpRequestFactory requestFactory = new JdkClientHttpRequestFactory();
-        requestFactory.setReadTimeout(Duration.ofMillis(requestTimeoutMs));
-        
-        this.restClient = RestClient.builder()
-            .requestFactory(requestFactory)
-            .build();
-    }
     
-    @Async("geoJsonExecutor")
-    public void sendGeoJson(GeoJsonObject geoJson) {
-        if (!geoJsonEnabled) {
-            return;
-        }
+    @Value("${geojson.jobIncludeList}")
+    private List<String> geoJsonJobIncludeList;
 
-        try {
-            String geoJsonString = objectMapper.writeValueAsString(geoJson);
-            restClient.post()
-                .uri(GEO_JSON_URL)
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(geoJsonString)
-                .retrieve();
-        } catch (JsonProcessingException e) {
-            log.error("Failed to send GeoJson", e);
+    @Autowired
+    private GeoJsonSenderService geoJsonService;
+
+    public void sendLineCrossings(List<LineCrossingObservation> lineCrossingObservations) {
+        // Filter observations for jobs that should be included in the GeoJSON
+        List<LineCrossingObservation> filteredObservations = lineCrossingObservations.stream()
+            .filter(lineCrossingObservation -> geoJsonJobIncludeList.contains(lineCrossingObservation.jobEntity().getName()))
+            .toList();
+
+        if (!filteredObservations.isEmpty()) {
+            geoJsonService.sendGeoJson(GeoJsonMapper.mapLineCrossings(filteredObservations));
+        }
+    }
+
+    public void sendAreaOccupancies(List<AreaOccupancyObservation> areaOccupancyObservations) {
+        // Filter observations for jobs that should be included in the GeoJSON
+        List<AreaOccupancyObservation> filteredObservations = areaOccupancyObservations.stream()
+            .filter(areaOccupancyObservation -> geoJsonJobIncludeList.contains(areaOccupancyObservation.jobEntity().getName()))
+            .toList();
+
+        if (!filteredObservations.isEmpty()) {
+            geoJsonService.sendGeoJson(GeoJsonMapper.mapAreaOccupancies(filteredObservations));
         }
     }
 }
